Renamed supporting Cellebrite classes to have consistent names
