+++ Experimental/src/org/sleuthkit/autopsy/experimental/cellex/datasourceprocessors/Bundle.properties
@@ -12,16 +12,16 @@ MPFModuleSimplePanel.jFilterFilesRadioButton.text=No
 MPFModuleSimplePanel.jLabel2.text=File Extensions:
 MPFModuleSimplePanel.jLabel3.text=Process files that meet the following criteria:
 MPFModuleSimplePanel.jLabel4.text=Run on all files in set of logical files?
-CellebriteXMLFilePanel.browseButton.text=Browse
-CellebriteXMLFilePanel.pathLabel.text=Browse for a Cellebrite XML file:
-CellebriteXMLFilePanel.pathTextField.text=
-CellebriteXMLFilePanel.jLabel1.text=Input type:
-CellebriteXMLFilePanel.jHandsetRadioButton.text=Handset
-CellebriteXMLFilePanel.jSIMRadioButton.text=SIM
-CellebriteAndroidInputPanel.browseButton.text=Browse
-CellebriteAndroidInputPanel.pathTextField.text=
-CellebriteAndroidInputPanel.timeZoneLabel.text=Please select the input timezone:
-CellebriteAndroidInputPanel.pathLabel.text=Browse for a Android phone image folder:
 OpenIDE-Module-Short-Description=Does triage of Android data.
-CellebriteAndroidInputPanel.errorLabel.text=Error Label
-CellebriteXMLFilePanel.errorLabel.text=Error Label
+CellebritePhysicalReportInputPanel.errorLabel.text=Error Label
+CellebritePhysicalReportInputPanel.timeZoneLabel.text=Please select the input timezone:
+CellebritePhysicalReportInputPanel.browseButton.text=Browse
+CellebritePhysicalReportInputPanel.pathTextField.text=
+CellebritePhysicalReportInputPanel.pathLabel.text=Browse for a Android phone image folder:
+CellebriteLogicalReportPanel.pathLabel.text=Browse for a Cellebrite XML file:
+CellebriteLogicalReportPanel.errorLabel.text=Error Label
+CellebriteLogicalReportPanel.jSIMRadioButton.text=SIM
+CellebriteLogicalReportPanel.jLabel1.text=Input type:
+CellebriteLogicalReportPanel.jHandsetRadioButton.text=Handset
+CellebriteLogicalReportPanel.browseButton.text=Browse
+CellebriteLogicalReportPanel.pathTextField.text=
