+++ Experimental/src/org/sleuthkit/autopsy/experimental/cellex/datasourceprocessors/CellebriteLogicalReportPanel.java
@@ -20,22 +20,21 @@ package org.sleuthkit.autopsy.experimental.cellex.datasourceprocessors;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
+import java.util.List;
 import java.io.File;
-import java.util.Calendar;
-import java.util.SimpleTimeZone;
-import java.util.TimeZone;
 import javax.swing.JFileChooser;
 import javax.swing.JPanel;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
+import javax.swing.filechooser.FileFilter;
 import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.autopsy.corecomponentinterfaces.DataSourceProcessor;
 import org.sleuthkit.autopsy.coreutils.ModuleSettings;
 import org.sleuthkit.autopsy.coreutils.PathValidator;
 
-public class CellebriteAndroidInputPanel extends JPanel implements DocumentListener {
+public class CellebriteLogicalReportPanel extends JPanel implements DocumentListener {
 
-    private final String PROP_LASTIMAGE_PATH = "LBL_LastImage_PATH";
+    private final String PROP_LASTINPUT_PATH = "LBL_LastInputFile_PATH";
 
     private PropertyChangeSupport pcs = null;
     private JFileChooser fc = new JFileChooser();
@@ -44,28 +43,37 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
     private String contextName;
 
     /**
-     * Creates new form CellebriteAndroidInputPanel
+     * Creates new form CellebriteXMLFilePanel
      */
-    public CellebriteAndroidInputPanel(String context) {
+    private CellebriteLogicalReportPanel(String context, List<FileFilter> fileChooserFilters) {
         initComponents();
 
         errorLabel.setVisible(false);
 
         fc.setDragEnabled(false);
-        fc.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
+        fc.setFileSelectionMode(JFileChooser.FILES_ONLY);
         fc.setMultiSelectionEnabled(false);
+
+        boolean firstFilter = true;
+        for (FileFilter filter : fileChooserFilters) {
+            if (firstFilter) {  // set the first on the list as the default selection
+                fc.setFileFilter(filter);
+                firstFilter = false;
+            } else {
+                fc.addChoosableFileFilter(filter);
+            }
+        }
 
         this.contextName = context;
         pcs = new PropertyChangeSupport(this);
-
-        createTimeZoneList();
     }
 
     /**
-     * Creates and returns an instance of a ImageFilePanel.
+     * Creates and returns an instance of a CellebriteLogicalReportPanel.
      */
-    public static synchronized CellebriteAndroidInputPanel createInstance(String context) {
-        CellebriteAndroidInputPanel instance = new CellebriteAndroidInputPanel(context);
+    public static synchronized CellebriteLogicalReportPanel createInstance(String context, List<FileFilter> fileChooserFilters) {
+        CellebriteLogicalReportPanel instance = new CellebriteLogicalReportPanel(context, fileChooserFilters);
+
         instance.postInit();
 
         return instance;
@@ -78,63 +86,6 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
     }
 
     /**
-     * Creates the drop down list for the time zones and then makes the local
-     * machine time zone to be selected.
-     */
-    public void createTimeZoneList() {
-        // load and add all timezone
-        String[] ids = SimpleTimeZone.getAvailableIDs();
-        for (String id : ids) {
-            TimeZone zone = TimeZone.getTimeZone(id);
-            int offset = zone.getRawOffset() / 1000;
-            int hour = offset / 3600;
-            int minutes = (offset % 3600) / 60;
-            String item = String.format("(GMT%+d:%02d) %s", hour, minutes, id);
-
-            /*
-             * DateFormat dfm = new SimpleDateFormat("z");
-             * dfm.setTimeZone(zone); boolean hasDaylight =
-             * zone.useDaylightTime(); String first = dfm.format(new Date(2010,
-             * 1, 1)); String second = dfm.format(new Date(2011, 6, 6)); int mid
-             * = hour * -1; String result = first + Integer.toString(mid);
-             * if(hasDaylight){ result = result + second; }
-             * timeZoneComboBox.addItem(item + " (" + result + ")");
-             */
-            timeZoneComboBox.addItem(item);
-        }
-        // get the current timezone
-        TimeZone thisTimeZone = Calendar.getInstance().getTimeZone();
-        int thisOffset = thisTimeZone.getRawOffset() / 1000;
-        int thisHour = thisOffset / 3600;
-        int thisMinutes = (thisOffset % 3600) / 60;
-        String formatted = String.format("(GMT%+d:%02d) %s", thisHour, thisMinutes, thisTimeZone.getID());
-
-        // set the selected timezone
-        timeZoneComboBox.setSelectedItem(formatted);
-    }
-
-    /**
-     * Get the path of the user selected folder.
-     *
-     * @return the image path
-     */
-    public String getContentPaths() {
-        return pathTextField.getText();
-    }
-
-    /**
-     * Set the path of the images folder.
-     */
-    public void setContentPath(String s) {
-        pathTextField.setText(s);
-    }
-
-    public String getTimeZone() {
-        String tz = timeZoneComboBox.getSelectedItem().toString();
-        return tz.substring(tz.indexOf(")") + 2).trim();
-    }
-
-    /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
@@ -143,30 +94,46 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
 
+        infileTypeButtonGroup = new javax.swing.ButtonGroup();
         pathLabel = new javax.swing.JLabel();
         pathTextField = new javax.swing.JTextField();
         browseButton = new javax.swing.JButton();
-        timeZoneLabel = new javax.swing.JLabel();
-        timeZoneComboBox = new javax.swing.JComboBox<>();
+        jHandsetRadioButton = new javax.swing.JRadioButton();
+        jLabel1 = new javax.swing.JLabel();
+        jSIMRadioButton = new javax.swing.JRadioButton();
         errorLabel = new javax.swing.JLabel();
 
-        org.openide.awt.Mnemonics.setLocalizedText(pathLabel, org.openide.util.NbBundle.getMessage(CellebriteAndroidInputPanel.class, "CellebriteAndroidInputPanel.pathLabel.text")); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(pathLabel, org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.pathLabel.text")); // NOI18N
 
-        pathTextField.setText(org.openide.util.NbBundle.getMessage(CellebriteAndroidInputPanel.class, "CellebriteAndroidInputPanel.pathTextField.text")); // NOI18N
+        pathTextField.setText(org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.pathTextField.text")); // NOI18N
 
-        org.openide.awt.Mnemonics.setLocalizedText(browseButton, org.openide.util.NbBundle.getMessage(CellebriteAndroidInputPanel.class, "CellebriteAndroidInputPanel.browseButton.text")); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(browseButton, org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.browseButton.text")); // NOI18N
         browseButton.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 browseButtonActionPerformed(evt);
             }
         });
 
-        org.openide.awt.Mnemonics.setLocalizedText(timeZoneLabel, org.openide.util.NbBundle.getMessage(CellebriteAndroidInputPanel.class, "CellebriteAndroidInputPanel.timeZoneLabel.text")); // NOI18N
+        infileTypeButtonGroup.add(jHandsetRadioButton);
+        org.openide.awt.Mnemonics.setLocalizedText(jHandsetRadioButton, org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.jHandsetRadioButton.text")); // NOI18N
+        jHandsetRadioButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jHandsetRadioButtonActionPerformed(evt);
+            }
+        });
 
-        timeZoneComboBox.setMaximumRowCount(30);
+        org.openide.awt.Mnemonics.setLocalizedText(jLabel1, org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.jLabel1.text")); // NOI18N
+
+        infileTypeButtonGroup.add(jSIMRadioButton);
+        org.openide.awt.Mnemonics.setLocalizedText(jSIMRadioButton, org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.jSIMRadioButton.text")); // NOI18N
+        jSIMRadioButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jSIMRadioButtonActionPerformed(evt);
+            }
+        });
 
         errorLabel.setForeground(new java.awt.Color(255, 0, 0));
-        org.openide.awt.Mnemonics.setLocalizedText(errorLabel, org.openide.util.NbBundle.getMessage(CellebriteAndroidInputPanel.class, "CellebriteAndroidInputPanel.errorLabel.text")); // NOI18N
+        org.openide.awt.Mnemonics.setLocalizedText(errorLabel, org.openide.util.NbBundle.getMessage(CellebriteLogicalReportPanel.class, "CellebriteLogicalReportPanel.errorLabel.text")); // NOI18N
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
@@ -175,15 +142,17 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(pathLabel)
                     .addGroup(layout.createSequentialGroup()
                         .addComponent(pathTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 286, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
                         .addComponent(browseButton))
+                    .addComponent(pathLabel)
                     .addGroup(layout.createSequentialGroup()
-                        .addComponent(timeZoneLabel)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(timeZoneComboBox, javax.swing.GroupLayout.PREFERRED_SIZE, 215, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(16, 16, 16)
+                        .addComponent(jHandsetRadioButton)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                        .addComponent(jSIMRadioButton))
+                    .addComponent(jLabel1)
                     .addComponent(errorLabel))
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
@@ -196,19 +165,20 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(pathTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(browseButton))
-                .addGap(8, 8, 8)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(errorLabel)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(timeZoneLabel)
-                    .addComponent(timeZoneComboBox, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(198, Short.MAX_VALUE))
+                .addComponent(jLabel1)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jSIMRadioButton)
+                    .addComponent(jHandsetRadioButton))
+                .addContainerGap(26, Short.MAX_VALUE))
         );
     }// </editor-fold>//GEN-END:initComponents
     @SuppressWarnings("deprecation")
     private void browseButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_browseButtonActionPerformed
       String oldText = pathTextField.getText();
-
       // set the current directory of the FileChooser if the ImagePath Field is valid
       File currentDir = new File(oldText);
       if (currentDir.exists()) {
@@ -220,18 +190,95 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
           String path = fc.getSelectedFile().getPath();
           pathTextField.setText(path);
       }
-
       pcs.firePropertyChange(DataSourceProcessor.DSP_PANEL_EVENT.FOCUS_NEXT.toString(), false, true);
     }//GEN-LAST:event_browseButtonActionPerformed
+
+    private void jHandsetRadioButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jHandsetRadioButtonActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jHandsetRadioButtonActionPerformed
+
+    private void jSIMRadioButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jSIMRadioButtonActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jSIMRadioButtonActionPerformed
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton browseButton;
     private javax.swing.JLabel errorLabel;
+    private javax.swing.ButtonGroup infileTypeButtonGroup;
+    private javax.swing.JRadioButton jHandsetRadioButton;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JRadioButton jSIMRadioButton;
     private javax.swing.JLabel pathLabel;
     private javax.swing.JTextField pathTextField;
-    private javax.swing.JComboBox<String> timeZoneComboBox;
-    private javax.swing.JLabel timeZoneLabel;
     // End of variables declaration//GEN-END:variables
+/**
+     * Get the path of the user selected image.
+     *
+     * @return the image path
+     */
+    public String getImageFilePath() {
+        return pathTextField.getText();
+    }
+
+    public void reset() {
+        //reset the UI elements to default 
+        pathTextField.setText(null);
+        jHandsetRadioButton.setSelected(true);
+    }
+
+    public boolean isHandsetFile() {
+        return jHandsetRadioButton.isSelected();
+    }
+
+    /**
+     * Should we enable the next button of the wizard?
+     *
+     * @return true if a proper image has been selected, false otherwise
+     */
+    public boolean validatePanel() {
+        errorLabel.setVisible(false);
+        String path = getImageFilePath();
+        if (path == null || path.isEmpty()) {
+            return false;
+        }
+
+        // display warning if there is one (but don't disable "next" button)
+        warnIfPathIsInvalid(path);
+
+        boolean isExist = new File(path).exists();
+
+        return (isExist);
+    }
+
+    /**
+     * Validates path to selected data source and displays warning if it is
+     * invalid.
+     *
+     * @param path Absolute path to the selected data source
+     */
+    private void warnIfPathIsInvalid(String path) {
+        if (!PathValidator.isValid(path, Case.getCurrentCase().getCaseType())) {
+            errorLabel.setVisible(true);
+            errorLabel.setText("Path to multi-user data source is on \"C:\" drive");
+        }
+    }
+
+    public void storeSettings() {
+        String inFilePath = getImageFilePath();
+        if (null != inFilePath) {
+            String imagePath = inFilePath.substring(0, inFilePath.lastIndexOf(File.separator) + 1);
+            ModuleSettings.setConfigSetting(contextName, PROP_LASTINPUT_PATH, imagePath);
+        }
+    }
+
+    public void readSettings() {
+        String inFilePath = ModuleSettings.getConfigSetting(contextName, PROP_LASTINPUT_PATH);
+        if (null != inFilePath) {
+            if (!inFilePath.isEmpty()) {
+                pathTextField.setText(inFilePath);
+            }
+        }
+    }
 
     /**
      * Update functions are called by the pathTextField which has this set as
@@ -278,69 +325,6 @@ public class CellebriteAndroidInputPanel extends JPanel implements DocumentListe
         super.removePropertyChangeListener(pcl);
 
         pcs.removePropertyChangeListener(pcl);
-    }
-
-    /**
-     * Should we enable the next button of the wizard?
-     *
-     * @return true if a proper image has been selected, false otherwise
-     */
-    public boolean validatePanel() {
-        errorLabel.setVisible(false);
-        String path = getContentPaths();
-        if (path == null || path.isEmpty()) {
-            return false;
-        }
-
-        // display warning if there is one (but don't disable "next" button)
-        warnIfPathIsInvalid(path);
-
-        // check if a folder exists by this name.
-        File f = new File(path);
-        boolean isValid = (f.exists() && f.isDirectory());
-
-        return isValid;
-    }
-
-    /**
-     * Validates path to selected data source and displays warning if it is
-     * invalid.
-     *
-     * @param path Absolute path to the selected data source
-     */
-    private void warnIfPathIsInvalid(String path) {
-        if (!PathValidator.isValid(path, Case.getCurrentCase().getCaseType())) {
-            errorLabel.setVisible(true);
-            errorLabel.setText("Path to multi-user data source is on \"C:\" drive");
-        }
-    }
-
-    public void reset() {
-        //reset the UI elements to default 
-        pathTextField.setText(null);
-    }
-
-    /*
-     * Store the last used settings
-     */
-    public void storeSettings() {
-        String imagePathName = getContentPaths();
-        if (null != imagePathName) {
-            String imagePath = imagePathName.substring(0, imagePathName.lastIndexOf(File.separator) + 1);
-            ModuleSettings.setConfigSetting(contextName, PROP_LASTIMAGE_PATH, imagePath);
-        }
-    }
-
-    /*
-     * Read and load the last used settings
-     */
-    public void readSettings() {
-        String lastImagePath = ModuleSettings.getConfigSetting(contextName, PROP_LASTIMAGE_PATH);
-        if (null != lastImagePath) {
-            if (!lastImagePath.isEmpty()) {
-                pathTextField.setText(lastImagePath);
-            }
-        }
     }
 
 }
